Fix: use BucketContext on textile crud operations
