+++ core/textile/client/client.go
@@ -1,16 +1,11 @@
 package client
 
 import (
-	"bytes"
 	"context"
 	"encoding/hex"
 	"errors"
-	"io"
 	"os"
-	"strings"
 	"time"
-
-	"github.com/ipfs/interface-go-ipfs-core/path"
 
 	buckets_pb "github.com/textileio/textile/api/buckets/pb"
 
@@ -328,49 +323,4 @@ func (tc *TextileClient) StartAndBootstrap() error {
 
 	log.Debug("Textile Client initialized successfully")
 	return nil
-}
-
-// UploadFile uploads a file to path on textile
-// path should include the file name as the last path segment
-// also nested path not existing yet would be created automatically
-func (tc *TextileClient) UploadFile(
-	ctx context.Context,
-	bucketKey string,
-	path string,
-	reader io.Reader,
-) (result path.Resolved, root path.Path, err error) {
-	return tc.buckets.PushPath(ctx, bucketKey, path, reader)
-}
-
-// CreateDirectory creates an empty directory
-// Because textile doesn't support empty directory an empty .keep file is created
-// in the directory
-func (tc *TextileClient) CreateDirectory(
-	ctx context.Context,
-	bucketKey string,
-	path string,
-) (result path.Resolved, root path.Path, err error) {
-	// append .keep file to the end of the directory
-	emptyDirPath := strings.TrimRight(path, "/") + "/" + keepFileName
-	return tc.buckets.PushPath(ctx, bucketKey, emptyDirPath, &bytes.Buffer{})
-}
-
-// ListDirectory returns a list of items in a particular directory
-func (tc *TextileClient) ListDirectory(
-	ctx context.Context,
-	bucketKey string,
-	path string,
-) (*TextileDirEntries, error) {
-	result, err := tc.buckets.ListPath(ctx, bucketKey, path)
-
-	return (*TextileDirEntries)(result), err
-}
-
-// DeleteDirOrFile will delete file or directory at path
-func (tc *TextileClient) DeleteDirOrFile(
-	ctx context.Context,
-	bucketKey string,
-	path string,
-) error {
-	return tc.buckets.RemovePath(ctx, bucketKey, path)
 }
